Fix comma-separated list on /learn.html.
The comma separated lists are really a display concern, and not a
markup concern - in particular, they're semantically lists.

In addition, laying them out like this means it's easy to forget
to put a trailing comma at the end of the line, like is present on
nixos.org/learn.html at the moment (which currently has:
"[Python], [Ruby] [Java], [Go], [Rust], "...).